+++ agent-control/src/agent_control/version_updater/on_host.rs
@@ -135,6 +135,7 @@ where
                 repository: self.repository.clone(),
                 version: new_version,
                 public_key_url: Some(self.pub_key_url.clone()),
+                postdownload: None,
             },
         }
     }