+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Remember that the keywords that you can use are the following:
 
 ## Unreleased
 
+### 🚀 Enhancements
+- enhancement: Add postdownload script support for OCI packages.
+
 ## v1.15.2 - 2026-05-29
 
 ### 🐞 Bug fixes